Move low-level SSH functions to ssh/client
