+++ pkg/ssh/client/ssh_suite_test.go
@@ -1,4 +1,4 @@
-package ssh
+package client
 
 import (
 	"testing"
